Tests added

+++ demo-server/src/main/java/com/illia/server/file_holder/FileHolderImpl.java
@@ -1,7 +1,6 @@
 package com.illia.server.file_holder;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -30,44 +29,52 @@ public class FileHolderImpl implements FileHolder {
 
     @Override
     public String saveFile(String fileName, MultipartFile mpFile) {
-        try {
-            var path = Path.of(prefix + fileName);
-            var updatedExisting = Files.deleteIfExists(path);
-            Files.createFile(path);
-            mpFile.transferTo(path);
-            savedFiles.put(fileName, path.toFile());
+        if (mpFile != null) {
+            try {
+                var path = Path.of(prefix + fileName);
+                var updatedExisting = Files.deleteIfExists(path);
+                Files.createFile(path);
+                mpFile.transferTo(path);
+                savedFiles.put(fileName, path.toFile());
 
-            if (updatedExisting) {
-                return "Updated existing file " + fileName;
-            } else {
-                return "Saved file " + fileName;
+                if (updatedExisting) {
+                    return "Updated existing file " + fileName;
+                } else {
+                    return "Saved file " + fileName;
+                }
+            } catch (Exception ex) {
+                ex.printStackTrace();
+                var internalErrorMsg = "Server error during 'save file' operation!";
+                log.error(internalErrorMsg);
+                return internalErrorMsg;
             }
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            var internalErrorMsg = "Server error during 'save file' operation!";
-            log.error(internalErrorMsg);
-            return internalErrorMsg;
+        } else {
+            return "File is null, nothing to save!";
         }
     }
 
     @Override
     public String saveFile(String fileName, File file) {
-        try {
-            var path = Path.of(prefix + fileName);
-            var updatedExisting = Files.deleteIfExists(path);
-            Files.move(file.toPath(), path);
-            savedFiles.put(fileName, path.toFile());
+        if (file != null) {
+            try {
+                var path = Path.of(prefix + fileName);
+                var updatedExisting = Files.deleteIfExists(path);
+                Files.move(file.toPath(), path);
+                savedFiles.put(fileName, path.toFile());
 
-            if (updatedExisting) {
-                return "Updated existing file " + fileName;
-            } else {
-                return "Saved file on server " + fileName;
+                if (updatedExisting) {
+                    return "Updated existing file " + fileName;
+                } else {
+                    return "Saved file on server " + fileName;
+                }
+            } catch (Exception ex) {
+                ex.printStackTrace();
+                var internalErrorMsg = "Server error during 'save file' operation!";
+                log.error(internalErrorMsg);
+                return internalErrorMsg;
             }
-        }catch (Exception ex){
-            ex.printStackTrace();
-            var internalErrorMsg = "Server error during 'save file' operation!";
-            log.error(internalErrorMsg);
-            return internalErrorMsg;
+        } else {
+            return "File is null, nothing to save!";
         }
     }
 
@@ -82,7 +89,7 @@ public class FileHolderImpl implements FileHolder {
     }
 
     @Override
-    public void setPrefix(String prefix){
+    public void setPrefix(String prefix) {
         this.prefix = prefix;
     }
 }
